feat: middlwares no laravel 11

+++ routes/web.php
@@ -2,15 +2,18 @@
 
 use App\Http\Controllers\Admin\UserController;
 use App\Http\Controllers\ProfileController;
+use App\Http\Middleware\CheckIfIsAdmin;
 use Illuminate\Support\Facades\Route;
 
-Route::delete('/users/{user}/destroy', [UserController::class, 'destroy'])->name('users.destroy');
-Route::get('/users/{user}', [UserController::class, 'show'])->name('users.show');
-Route::put('/users/{user}', [UserController::class, 'update'])->name('users.update');
-Route::post('/users', [UserController::class, 'store'])->name('users.store');
-Route::get('/users/{user}/edit', [UserController::class, 'edit'])->name('users.edit');
-Route::get('/users/create', [UserController::class, 'create'])->name('users.create');
-Route::get('/users', [UserController::class, 'index'])->name('users.index');
+Route::middleware('auth')->group( function(){
+    Route::delete('/users/{user}/destroy', [UserController::class, 'destroy'])->name('users.destroy')->middleware(CheckIfIsAdmin::class);
+    Route::get('/users/{user}', [UserController::class, 'show'])->name('users.show');
+    Route::put('/users/{user}', [UserController::class, 'update'])->name('users.update');
+    Route::post('/users', [UserController::class, 'store'])->name('users.store');
+    Route::get('/users/{user}/edit', [UserController::class, 'edit'])->name('users.edit');
+    Route::get('/users/create', [UserController::class, 'create'])->name('users.create');
+    Route::get('/users', [UserController::class, 'index'])->name('users.index');
+});
 
 Route::get('/', function () {
     return view('welcome');
